refactor: move logic out of `addons_manager`
